BuyerView, ArticlePreview, modifica Article con image

+++ DietiDeals24/ReverseAuctionModel.swift
@@ -9,6 +9,6 @@ import Foundation
 
 class ReverseAuction: Auction {
     var decrease: Float?
-    var decreaseTimer: Float? = 6000
+    var decreaseTimer: Int? = 6000
     var minimumPrice: Float?
 }
